+++ TurretGame/AreaEffect.cpp
@@ -6,8 +6,8 @@ void AreaEffect::Update(unsigned int frame)
 
 void AreaEffect::Draw()
 {
-	DrawCircleLines(this->position.x, this->position.y, 30.0f, RED);
-	DrawText("AOE HERE", this->position.x, this->position.y, 20.0f, RED);
+	DrawCircleLines((int)this->position.x, (int)this->position.y, 30.0f, RED);
+	DrawText("AOE HERE", (int)this->position.x, (int)this->position.y, 20, RED);
 }
 
 int AreaEffect::GetID()
