+++ TurretGame/SoldierEnemy.cpp
@@ -26,7 +26,7 @@ void SoldierEnemy::Update(unsigned int frame)
 	Enemy::Update(frame);
 
 	//update position of texture
-	this->textureLoop->SetPosition(this->position.x, this->position.y);
+	this->textureLoop->SetPosition((int)this->position.x, (int)this->position.y);
 
 	//update the animation state
 	if (frame % 6 == 0) this->textureLoop->NextFrame();
